% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{binarizeMatrix}
\alias{binarizeMatrix}
\title{Binarize a matrix}
\usage{
binarizeMatrix(A)
}
\arguments{
\item{A}{matrix}
}
\value{
Matrix binarized
}
\description{
Get a binarized version (only 0s and 1s) of a Matrix, which is assumed to be
sparse. Matrix entries which are different from zero are set to one.
}
