% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{blkdiag}
\alias{blkdiag}
\title{Return a list of matrices in block diagonal form}
\usage{
blkdiag(...)
}
\arguments{
\item{...}{individual matrices or a list of matrices}
}
\description{
This function is a wrapper for \link[Matrix]{bdiag}
}
\seealso{
\link[Matrix]{bdiag}
}
