% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{BuildDensityMatrixBGS}
\alias{BuildDensityMatrixBGS}
\title{Calculate the density matrix from an adjacency matrix}
\usage{
BuildDensityMatrixBGS(AdjacencyMatrix)
}
\arguments{
\item{AdjacencyMatrix}{an adjacency matrix}
}
\value{
Density matrix for the provided matrix
}
\description{
Calculate the density matrix from an adjacency matrix
}
\references{
S. L. Braunstein, S. Ghosh, S. Severini, Annals of Combinatorics 10, No 3, (2006)

De Domenico, M., Set al. (2013). Mathematical formulation of
multilayer networks. Physical Review X, 3(4), 041022.
\href{https://doi.org/10.1103/PhysRevX.3.041022}{doi 10.1103/PhysRevX.3.041022}
}
