% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{BuildExtendedEdgelistFromSupraAdjacencyMatrix}
\alias{BuildExtendedEdgelistFromSupraAdjacencyMatrix}
\title{Builds extended edgelist from supra-adjacency matrix}
\usage{
BuildExtendedEdgelistFromSupraAdjacencyMatrix(
  SupraAdjacencyMatrix,
  Layers,
  Nodes,
  isDirected
)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}

\item{isDirected}{logical}
}
\value{
Data frame with columns: \code{node.from, layer.from, node.to, layer.to weight}
}
\description{
Builds extended edgelist from supra-adjacency matrix
}
