% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{BuildLayersTensor}
\alias{BuildLayersTensor}
\title{Build layers tensor}
\usage{
BuildLayersTensor(Layers, OmegaParameter, MultisliceType)
}
\arguments{
\item{Layers}{scalar, number of layers, number of layers}

\item{OmegaParameter}{scalar, weight of links (future inter-layer links)}

\item{MultisliceType}{"ordered": chain, undirected; "categorical": all-to-all, undirected; "temporal": chain, directed}
}
\value{
Tensor matrix.
}
\description{
Build layers tensor
}
