% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{buildMultilayerNetworkFromMuxvizFiles}
\alias{buildMultilayerNetworkFromMuxvizFiles}
\title{Return a Multilayer Network from Muxviz files}
\usage{
buildMultilayerNetworkFromMuxvizFiles(
  config.file,
  isDirected,
  isWeighted,
  MultisliceType,
  LayerCouplingStrength = 1,
  format = "muxviz edge-colored",
  verbose = T
)
}
\arguments{
\item{config.file}{path, to the config file}

\item{isDirected}{logical, for considering directed networks}

\item{isWeighted}{logical, for considering weighted networks}

\item{MultisliceType}{string, "ordinal" or "categorical" or "temporal"}

\item{LayerCouplingStrength}{used only if format is muxviz edge-colored}

\item{format}{string, "muxviz edge-colored" or "muxviz general"}

\item{verbose}{logical, for messages about processing}
}
\value{
A list including nodes, layers and the graph features
}
\description{
Return a Multilayer Network from Muxviz files
}
