% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{BuildSupraAdjacencyMatrixFromEdgeColoredMatrices}
\alias{BuildSupraAdjacencyMatrixFromEdgeColoredMatrices}
\title{Build supra-adjacency matrix from edge-colored matrices}
\usage{
BuildSupraAdjacencyMatrixFromEdgeColoredMatrices(
  NodesTensor,
  LayerTensor,
  Layers,
  Nodes
)
}
\arguments{
\item{NodesTensor}{list of adjacency matrices, expected to be aligned (a node hassame index in any layer)}

\item{LayerTensor}{matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
a list of matrices. Let us call \code{AdjMatrix} the output of the function,
AdjMatrix (that here would play the role of NodesTensor) is vectorlist with
\code{Layers + 1} entries. Use \code{AdjMatrix[1:Layers]} to obtain the expected result.
}
\description{
Build supra-adjacency matrix from edge-colored matrices
}
