% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{BuildSupraAdjacencyMatrixFromExtendedEdgelist}
\alias{BuildSupraAdjacencyMatrixFromExtendedEdgelist}
\title{Build supra-adjacency matrix from edge lists}
\usage{
BuildSupraAdjacencyMatrixFromExtendedEdgelist(
  mEdges,
  Layers,
  Nodes,
  isDirected
)
}
\arguments{
\item{mEdges}{data frame with extended edge list, i.e a data frame with
columns: \code{node.from, layer.from, node.to, layer.to weight}}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}

\item{isDirected}{logical}
}
\value{
Supra-adjacency matrix, a square matrix of dimension \code{Nodes * Layers}.
}
\description{
Build supra-adjacency matrix from edge lists
}
