% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{BuildSupraTransitionMatrixFromSupraAdjacencyMatrix}
\alias{BuildSupraTransitionMatrixFromSupraAdjacencyMatrix}
\title{Build the supra-transition matrix for a multilayer network}
\usage{
BuildSupraTransitionMatrixFromSupraAdjacencyMatrix(
  SupraAdjacencyMatrix,
  Layers,
  Nodes,
  Type = "pagerank",
  r = NULL
)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adjacency matrix, which contains
the information in intra-layer connectivity \eqn{\{W^{\alpha}_{ij}\}} and
on inter-layer coupling \eqn{D^{\alpha \beta}_{i}\}_i}.}

\item{Layers}{integer, number of layers}

\item{Nodes}{integer, number of nodes}

\item{Type}{the type of the random walk, default \code{"pagerank"}.
Other types:
\code{c("classical", "diffusive", "maxent", "physical", "relaxed-physical")}.
For details on the transition rules see Refs.2-3.}

\item{r}{relaxation parameter, should be a float in the range \eqn{[0, 1]}.
Default is NULL, which implies \eqn{r=0.5} for the physical random
walk with relaxation (\code{Type = "relaxed-physical"}).}
}
\value{
the supra-transition matrix of probabilities for a
discrete-time random walk on the given multi-layer network.
}
\description{
Given a supra-adjacency matrix, build supra-transition the
matrix corresponding to different type of random walks (discrete time
Markov chians).
The types of RW considered here are the same as in Ref.2-Supplementary Material: classical, diffusive, PageRank, physical, physical with relaxation
and maximum-entropy.

The maximum entropy random walk (MERW) choses the stochastic matrix which
maximises \eqn{H(S)}, so that the walker can explore every walk of the
same length with equal probability.
Let \eqn{\lambda_N, \phi} be the leading eigenvalue and
corresponding right eigenvector of the adjacency matrix \eqn{A}. Then
\eqn{T_{ij} = \frac{A_{ij}}{\lambda_N}\frac{\phi_j}{\phi_i}.}
}
\references{
\enumerate{
\item De Domenico, M., Set al. (2013). Mathematical formulation of
multilayer networks. Physical Review X, 3(4), 041022.
\href{https://doi.org/10.1103/PhysRevX.3.041022}{doi 10.1103/PhysRevX.3.041022}
\item De Domenico, M., et al. (2014). Navigability of interconnected networks
under random failures. PNAS, 111(23).
\href{https://doi.org/10.1073/pnas.1318469111}{doi 10.1073/pnas.1318469111}
\item Bertagnolli, G., & De Domenico, M. (2020). Diffusion Geometry of
Multiplex and Interdependent Systems.
\href{http://arxiv.org/abs/2006.13032}{arxiv preprint arxiv:2006.13032}
}
}
