% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{BuildTransitionMatrixFromEdgeColoredMatrices}
\alias{BuildTransitionMatrixFromEdgeColoredMatrices}
\title{Random walk transition matrix for an edge-coloured network.
Given the node tensor for the edge-coloured multi-layer network
the function builds the supra transition matrix for a discrete time RW.}
\usage{
BuildTransitionMatrixFromEdgeColoredMatrices(
  NodesTensor,
  Layers,
  Nodes,
  Type = "classical"
)
}
\arguments{
\item{NodesTensor}{list of adjacency matrices, expected to be aligned (a node
has same index in any layer).}

\item{Layers}{integer, number of layers}

\item{Nodes}{integer, number of nodes}

\item{Type}{the type of the random walk, default \code{"classical"}.
Other types: \code{c("pagerank", "diffusive", "maxent")}.
For details on the transition rules see Refs.2-3.}
}
\value{
a square matrix of dimension \code{Nodes} of transition probabilities for a
discrete-time random walk on the given edge-coloured network.
}
\description{
Random walk transition matrix for an edge-coloured network.
Given the node tensor for the edge-coloured multi-layer network
the function builds the supra transition matrix for a discrete time RW.
}
\references{
\enumerate{
\item De Domenico, M., Set al. (2013). Mathematical formulation of
multilayer networks. Physical Review X, 3(4), 041022.
\href{https://doi.org/10.1103/PhysRevX.3.041022}{doi 10.1103/PhysRevX.3.041022}
\item De Domenico, M., et al. (2014). Navigability of interconnected networks
under random failures. PNAS, 111(23).
\href{https://doi.org/10.1073/pnas.1318469111}{doi 10.1073/pnas.1318469111}
\item Bertagnolli, G., & De Domenico, M. (2020). Diffusion Geometry of
Multiplex and Interdependent Systems.
\href{http://arxiv.org/abs/2006.13032}{arxiv preprint arxiv:2006.13032}
}
}
