% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{BuildTransitionMatrixFromSingleLayer}
\alias{BuildTransitionMatrixFromSingleLayer}
\title{Returns a Transition Matrix from a single Layer}
\usage{
BuildTransitionMatrixFromSingleLayer(AdjacencyMatrix, Nodes)
}
\arguments{
\item{AdjacencyMatrix}{matrix}

\item{Nodes}{scalar, number of nodes}
}
\value{
Transition Matrix
}
\description{
Returns a Transition Matrix from a single Layer
}
