% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{CanonicalVector}
\alias{CanonicalVector}
\title{Return the ith canonical vector of a N-dimesnion basis}
\usage{
CanonicalVector(N, i)
}
\arguments{
\item{N}{sacalar, dimension of the basis}

\item{i}{sacalar, the ith vector}
}
\value{
Canonical vector
}
\description{
Return the ith canonical vector of a N-dimesnion basis
}
