% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{diagR}
\alias{diagR}
\title{Return a matrix with elements on a diagonal specified by offset}
\usage{
diagR(x, n, offset = 0)
}
\arguments{
\item{x}{fill diagonal with this value}

\item{n}{matrix size}

\item{offset}{start position of diagonal, default: 0 = main diagonal}
}
\value{
A diagonal matrix with \code{x} on its diagonal
(main or offset diagonal).
}
\description{
Return a matrix with elements on a diagonal specified by offset
}
