% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetAggregateMatrix}
\alias{GetAggregateMatrix}
\title{Get aggregate matrix from nodes tensor}
\usage{
GetAggregateMatrix(NodesTensor, Layers, Nodes)
}
\arguments{
\item{NodesTensor}{list of adjacency matrices, expected to be aligned (a node has same index in any layer)}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
Aggregate matrix
}
\description{
Get aggregate matrix from nodes tensor
}
