% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetAggregateMatrixFromNetworkList}
\alias{GetAggregateMatrixFromNetworkList}
\title{Get aggregate Matrix from network List}
\usage{
GetAggregateMatrixFromNetworkList(g.list)
}
\arguments{
\item{g.list}{list of igraph objects}
}
\value{
Aggregate matrix
}
\description{
Get aggregate Matrix from network List
}
