% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetAggregateNetworkFromNetworkList}
\alias{GetAggregateNetworkFromNetworkList}
\title{Get aggregate Network from network List}
\usage{
GetAggregateNetworkFromNetworkList(g.list)
}
\arguments{
\item{g.list}{list of igraph objects}
}
\value{
Aggregate network
}
\description{
Get aggregate Network from network List
}
