% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetAggregateNetworkFromSupraAdjacencyMatrix}
\alias{GetAggregateNetworkFromSupraAdjacencyMatrix}
\title{Get aggregate network from supra-adjacency matrix}
\usage{
GetAggregateNetworkFromSupraAdjacencyMatrix(
  SupraAdjacencyMatrix,
  Layers,
  Nodes
)
}
\arguments{
\item{SupraAdjacencyMatrix}{supra-adjacency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
Aggregate network
}
\description{
Get aggregate network from supra-adjacency matrix
}
