% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetAverageGlobalNodeOverlappingMatrix}
\alias{GetAverageGlobalNodeOverlappingMatrix}
\title{Calculates the average global node overlap Matrix}
\usage{
GetAverageGlobalNodeOverlappingMatrix(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
Average Global Node Overlapping Matrix
}
\description{
Calculates the average global node overlap Matrix
}
\references{
De Domenico, M., et al. Structural reducibility of multilayer networks. Nat Commun 6, 6864 (2015).
\href{https://doi.org/10.1038/ncomms7864}{doi 10.1038/ncomms7864}
}
