% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetAverageGlobalOverlappingMatrix}
\alias{GetAverageGlobalOverlappingMatrix}
\title{Calculates the average global edge overlap Matrix}
\usage{
GetAverageGlobalOverlappingMatrix(
  SupraAdjacencyMatrix,
  Layers,
  Nodes,
  fastBinary = F
)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}

\item{fastBinary}{logical, optimization for undirected, unweighted, networks}
}
\value{
Average Global Edge Overlapping Matrix
}
\description{
Calculates the average global edge overlap Matrix
}
\references{
De Domenico, M., et al. Structural reducibility of multilayer networks. Nat Commun 6, 6864 (2015).
\href{https://doi.org/10.1038/ncomms7864}{doi 10.1038/ncomms7864}
}
