% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetConfigurationModelTypeIIFromNetworkList}
\alias{GetConfigurationModelTypeIIFromNetworkList}
\title{Generate random multiplex networks with fixed degree sequences, while destroying both intra- and inter-layer degree correlations}
\usage{
GetConfigurationModelTypeIIFromNetworkList(
  g.list,
  isDirected = NULL,
  method = "vl",
  verbose = F
)
}
\arguments{
\item{g.list}{list of igraph network objects}

\item{isDirected}{specifies if the networks are directed or not. Default: NULL (automatic selection)}

\item{method}{specifies which igraph method to use. Default: vl (slower but more accurate), other options: "simple", "simple.no.multiple"}

\item{verbose}{logical Default: F (automatic selection)}
}
\value{
A list of igraph network objects
}
\description{
Generate random multiplex networks with fixed degree sequences, while destroying both intra- and inter-layer degree correlations
}
