% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetConnectedComponents}
\alias{GetConnectedComponents}
\title{Return the component assignment for each node}
\usage{
GetConnectedComponents(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
List of Connected Components
}
\description{
Return the component assignment for each node
}
