% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetCoverageEvolutionEdgeColored}
\alias{GetCoverageEvolutionEdgeColored}
\title{Get the Coverage Evolution of an edge-colored network}
\usage{
GetCoverageEvolutionEdgeColored(
  TransitionMatrix,
  Layers,
  Nodes,
  TimeSequence,
  Approximate = FALSE,
  Approximate.disconnected = 0
)
}
\arguments{
\item{TransitionMatrix}{the non-interconnected multiplex transition matrix of
an edge-colored network, as returned from
\link{BuildTransitionMatrixFromEdgeColoredMatrices}}

\item{Layers}{Layers}

\item{Nodes}{Layers}

\item{TimeSequence}{Time sequence}

\item{Approximate}{logical, use approximate method. Default FALSE}

\item{Approximate.disconnected}{Numeric for approximate method. Default 0}
}
\value{
data frame coverage \code{rho} across time \code{tau}
}
\description{
This function calls \link{GetCoverageEvolutionMultilayer} with \code{Layers = 1}.
}
