% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetCoverageEvolutionMultilayer}
\alias{GetCoverageEvolutionMultilayer}
\title{Get the Coverage Evolution of a Multilayer Network}
\usage{
GetCoverageEvolutionMultilayer(
  SupraTransitionMatrix,
  Layers,
  Nodes,
  TimeSequence,
  Approximate = FALSE,
  Approximate.disconnected = 0
)
}
\arguments{
\item{SupraTransitionMatrix}{a Supratransition Matrix}

\item{Layers}{Layers}

\item{Nodes}{Layers}

\item{TimeSequence}{Time sequence}

\item{Approximate}{logical, use approximate method}

\item{Approximate.disconnected}{Numeric for approximate method}
}
\value{
Kronecker product of A and B
}
\description{
Get the Coverage Evolution of a Multilayer Network
}
\references{
M. De Domenico, A. Sole-Ribalta, S. Gomez, A. Arenas, PNAS 11, 8351 (2014)
}
