% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetEigenvaluesOfDensityMatrix}
\alias{GetEigenvaluesOfDensityMatrix}
\title{Calculate the eigenvalues of a density matrix}
\usage{
GetEigenvaluesOfDensityMatrix(DensityMatrix)
}
\arguments{
\item{DensityMatrix}{an adjacency matrix}
}
\value{
The Eigenvalues for the provided Density matrix
}
\description{
Calculate the eigenvalues of a density matrix
}
\references{
De Domenico, M., Set al. (2013). Mathematical formulation of
multilayer networks. Physical Review X, 3(4), 041022.
\href{https://doi.org/10.1103/PhysRevX.3.041022}{doi 10.1103/PhysRevX.3.041022}
}
