% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetEigenvaluesOfDensityMatrixFromAdjacencyMatrix}
\alias{GetEigenvaluesOfDensityMatrixFromAdjacencyMatrix}
\title{Calculate the eigenvalues of a density matrix from an adjacency matrix}
\usage{
GetEigenvaluesOfDensityMatrixFromAdjacencyMatrix(AdjacencyMatrix)
}
\arguments{
\item{AdjacencyMatrix}{an adjacency matrix}
}
\value{
The Eigenvalues for the Density matrix of the provided adjacency matrix
}
\description{
Calculate the eigenvalues of a density matrix from an adjacency matrix
}
