% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetGiantConnectedComponent}
\alias{GetGiantConnectedComponent}
\title{Return the largest connected component}
\usage{
GetGiantConnectedComponent(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
Largest Conected Component
}
\description{
Return the largest connected component
}
