% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetGiantIntersectionComponent}
\alias{GetGiantIntersectionComponent}
\title{Return the largest connected component obtained by intersecting LCC of each layer. Assumption: the network is an edge-colored multigraph}
\usage{
GetGiantIntersectionComponent(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
The largest connected component
}
\description{
Return the largest connected component obtained by intersecting LCC of each layer. Assumption: the network is an edge-colored multigraph
}
