% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetGiantIntersectionComponentFromNetworkList}
\alias{GetGiantIntersectionComponentFromNetworkList}
\title{Return the largest connected component obtained by intersecting LCC of each layer. Assumption: the network is an edge-colored multigraph}
\usage{
GetGiantIntersectionComponentFromNetworkList(g.list)
}
\arguments{
\item{g.list}{list of igraph network objects}
}
\value{
The largest connected component
}
\description{
Return the largest connected component obtained by intersecting LCC of each layer. Assumption: the network is an edge-colored multigraph
}
