% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetGiantViableComponent}
\alias{GetGiantViableComponent}
\title{Return the largest viable cluster}
\usage{
GetGiantViableComponent(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
The largest viable cluster
}
\description{
Return the largest viable cluster
}
\references{
Baxter et al, Physical Review Letters 109, 248701 (2012)
10.1103/PhysRevLett.109.248701
}
