% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetGiantViableComponentFromNetworkList}
\alias{GetGiantViableComponentFromNetworkList}
\title{Return the largest viable cluster from network list. Assumption: the network is an edge-colored multigraph}
\usage{
GetGiantViableComponentFromNetworkList(g.list)
}
\arguments{
\item{g.list}{list of igraph network objects}
}
\value{
The largest viable cluster
}
\description{
Return the largest viable cluster from network list. Assumption: the network is an edge-colored multigraph
}
\references{
Baxter et al, Physical Review Letters 109, 248701 (2012)
10.1103/PhysRevLett.109.248701
}
