% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetJensenShannonDivergence}
\alias{GetJensenShannonDivergence}
\title{Calculate the Jensen-Shannon Divergence of two networks}
\usage{
GetJensenShannonDivergence(
  AdjacencyMatrix1,
  AdjacencyMatrix2,
  VNEntropy1,
  VNEntropy2
)
}
\arguments{
\item{AdjacencyMatrix1}{an adjacency matrix}

\item{AdjacencyMatrix2}{an adjacency matrix}

\item{VNEntropy1}{the value of the Von Neumann entropy of the corresponding (1st) adjacency matrix}

\item{VNEntropy2}{the value of the Von Neumann entropy of the corresponding (2nd) adjacency matrix}
}
\value{
The Jensen–Shannon distance between the two networks
}
\description{
Calculate the Jensen-Shannon Divergence of two networks
}
\references{
De Domenico, M., et al. Structural reducibility of multilayer networks. Nat Commun 6, 6864 (2015).
\href{https://doi.org/10.1038/ncomms7864}{doi 10.1038/ncomms7864}
}
