% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetLaplacianMatrix}
\alias{GetLaplacianMatrix}
\title{Network combinatorial Laplacian}
\usage{
GetLaplacianMatrix(AdjacencyMatrix)
}
\arguments{
\item{AdjacencyMatrix}{the adjacency matrix characterising the network}
}
\value{
The network combinatorial Lapalacian Matrix
}
\description{
Given an adjacency matrix \eqn{A}, the function builds the
combinatorial Laplacian \eqn{D - A} for a single-layer network.
}
