% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetLargestEigenv}
\alias{GetLargestEigenv}
\title{Gets the largest Eigenvalue from a Matrix. A warning is raised if complex numbers emerge.}
\usage{
GetLargestEigenv(Matrix)
}
\arguments{
\item{Matrix}{matrix}
}
\value{
The eigenvector of largest eigenvalue, the largest eigenvalue
}
\description{
Gets the largest Eigenvalue from a Matrix. A warning is raised if complex numbers emerge.
}
