% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetLocalClustering}
\alias{GetLocalClustering}
\title{Calculate Local Clustering Coefficient}
\usage{
GetLocalClustering(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
Local clustering coefficient
}
\description{
Calculate Local Clustering Coefficient
}
\references{
De Domenico, M., Set al. (2013). Mathematical formulation of
multilayer networks. Physical Review X, 3(4), 041022.
\href{https://doi.org/10.1103/PhysRevX.3.041022}{doi 10.1103/PhysRevX.3.041022}
}
