% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMonoplexTriads}
\alias{GetMonoplexTriads}
\title{Return the closed triangles from a monoplex network}
\usage{
GetMonoplexTriads(g)
}
\arguments{
\item{g}{graph}
}
\value{
A dataframe containing the triangles of the network
}
\description{
Return the closed triangles from a monoplex network
}
\references{
De Domenico, M., Set al. (2013). Mathematical formulation of
multilayer networks. Physical Review X, 3(4), 041022.
\href{https://doi.org/10.1103/PhysRevX.3.041022}{doi 10.1103/PhysRevX.3.041022}

Cozzo, E.,Kivela, M., De Domenico, M., et al, New J. of Phys. 17, 073029 (2015)
\href{https://doi.org/10.1088/1367-2630/17/7/073029}{doi 10.1088/1367-2630/17/7/073029}
}
