% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultiClosenessCentrality}
\alias{GetMultiClosenessCentrality}
\title{Returns Multi Closeness Centrality}
\usage{
GetMultiClosenessCentrality(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
Vector of Closeness centralities
}
\description{
Returns Multi Closeness Centrality
}
