% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultiEigenvectorCentrality}
\alias{GetMultiEigenvectorCentrality}
\title{Calculates Multi-Eigenvector-Centrality}
\usage{
GetMultiEigenvectorCentrality(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
Vector of Eigenvector centralities
}
\description{
Calculates Multi-Eigenvector-Centrality
}
\references{
De Domenico, M., Solé-Ribalta, A., Omodei, E. et al. Ranking in interconnected multilayer
networks reveals versatile nodes. Nat Commun 6, 6868 (2015).
\href{https://doi.org/10.1038/ncomms7868}{doi 10.1038/ncomms7868}
}
