% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultiKCoreCentrality}
\alias{GetMultiKCoreCentrality}
\title{Calculates Multi-KCore Centrality. Computes centrality in each layer separately and then get the max per node.}
\usage{
GetMultiKCoreCentrality(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
Vector of K-core centralities
}
\description{
Calculates Multi-KCore Centrality. Computes centrality in each layer separately and then get the max per node.
}
