% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultilayerMotifs}
\alias{GetMultilayerMotifs}
\title{Return the multilayer motifs.
Note that even if inter-links are present, they will be neglected and the network treated as an edge-colored multigraph}
\usage{
GetMultilayerMotifs(
  SupraAdjacencyMatrix,
  Layers,
  Nodes,
  bin.path = NA,
  motifSize = 3,
  motifSamples = 1e+05,
  isDirected = FALSE,
  motifNullModel = "Local const",
  randomNetworks = 1000,
  randomExchangePerEdges = 3,
  randomExchangeAttempts = 3
)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}

\item{bin.path}{path, full path where the FANMOD binary executable is located}

\item{motifSize}{integer, size of motifs to find (3 or 4)}

\item{motifSamples}{integer, number of samples used to determine approx. # of motifs}

\item{isDirected}{logical, indicates if the search is for an undirected or directed network}

\item{motifNullModel}{string, "Local const", "Global const", or "No regard"}

\item{randomNetworks}{integer, number of random networks}

\item{randomExchangePerEdges}{integer, number of exchanges per edge}

\item{randomExchangeAttempts}{integer, number of exchange attempts per edge}
}
\value{
Motifs Table
}
\description{
Return the multilayer motifs.
Note that even if inter-links are present, they will be neglected and the network treated as an edge-colored multigraph
}
