% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultilayerMotifsFromNetworkList}
\alias{GetMultilayerMotifsFromNetworkList}
\title{Calculates the Motifs of a Multilayer from a network list}
\usage{
GetMultilayerMotifsFromNetworkList(
  g.list,
  bin.path = NA,
  motifSize = 3,
  motifSamples = 1e+05,
  isDirected = FALSE,
  motifNullModel = "Local const",
  randomNetworks = 1000,
  randomExchangePerEdges = 3,
  randomExchangeAttempts = 3
)
}
\arguments{
\item{g.list}{list of igraph network objects}

\item{bin.path}{path, full path where the FANMOD binary executable is located}

\item{motifSize}{integer, size of motifs to find (3 or 4)}

\item{motifSamples}{integer, number of samples used to determine approx. # of motifs}

\item{isDirected}{logical, indicates if the search is for an undirected or directed network}

\item{motifNullModel}{string, "Local const", "Global const", or "No regard"}

\item{randomNetworks}{integer, number of random networks}

\item{randomExchangePerEdges}{integer, number of exchanges per edge}

\item{randomExchangeAttempts}{integer, number of exchange attempts per edge}
}
\value{
Motifs Table
}
\description{
Calculates the Motifs of a Multilayer from a network list
}
\references{
:
Sebastian Wernicke and Florian Rasche, Bioinformatics 22, 1152 (2006)
\href{https://doi.org/10.1093/bioinformatics/btl038}{doi 10.1093/bioinformatics/btl038}

Version implemented by Tomer Benyamini, Yoav Teboulle, Tel-Aviv University
\href{github/gtremper/Network-Motif}{https://github.com/gtremper/Network-Motif/tree/master/fanmod/FANMOD-command_line-source}
}
