% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultilayerReducibility}
\alias{GetMultilayerReducibility}
\title{Calculate the Reducibility of a multilayer network}
\usage{
GetMultilayerReducibility(SupraAdjacencyMatrix, Layers, Nodes, Method, Type)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}

\item{Method}{string}

\item{Type}{string,  "Ordinal" or "Categorical"}
}
\value{
A list which element are: the Jensen–Shannon distance, the global quality function, and the relative quality function
}
\description{
Calculate the Reducibility of a multilayer network
}
\references{
De Domenico, M., et al. Structural reducibility of multilayer networks. Nat Commun 6, 6864 (2015).
\href{https://doi.org/10.1038/ncomms7864}{doi 10.1038/ncomms7864}

De Domenico, M., Set al. (2013). Mathematical formulation of
multilayer networks. Physical Review X, 3(4), 041022.
\href{https://doi.org/10.1103/PhysRevX.3.041022}{doi 10.1103/PhysRevX.3.041022}
}
