% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultiPageRankCentrality}
\alias{GetMultiPageRankCentrality}
\title{Calculates Multi-PageRank-Centrality (based on Multi-RandomWalk-Centrality)}
\usage{
GetMultiPageRankCentrality(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
Vector of PageRank centralities
}
\description{
Calculates Multi-PageRank-Centrality (based on Multi-RandomWalk-Centrality)
}
\references{
De Domenico, M., Solé-Ribalta, A., Omodei, E. et al. Ranking in interconnected multilayer
networks reveals versatile nodes. Nat Commun 6, 6868 (2015).
\href{https://doi.org/10.1038/ncomms7868}{doi 10.1038/ncomms7868}
}
