% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultiPathStatistics}
\alias{GetMultiPathStatistics}
\title{Return statistics based on multilayer paths}
\usage{
GetMultiPathStatistics(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency sparse matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
A a list containing the shortest-path distance matrix, the average path length, and the closeness versatility
}
\description{
Return statistics based on multilayer paths
}
