% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultiplexCommunities_Infomap}
\alias{GetMultiplexCommunities_Infomap}
\title{Return multilayer communities. INFOMAP wrapper}
\usage{
GetMultiplexCommunities_Infomap(
  g.list,
  bin.path = NA,
  isDirected,
  seed = 12345,
  includeSelfLinks = F,
  numTrials = 100,
  twoLevel = T,
  preclusterMultiplex = F,
  addMissingPhysicalNodes = T,
  hardPartitions = F,
  verbose = T,
  addAggregateAnalysis = T,
  multilayerRelaxRate = NA,
  multilayerJSRelaxRate = NA,
  outputPrefix = "multimap"
)
}
\arguments{
\item{g.list}{list of igraph network objects}

\item{bin.path}{path, full path where the INFOMAP binary executable is located}

\item{isDirected}{logical, indicates if the analysis is for an undirected or directed network}

\item{seed}{Integer, Infomap seed}

\item{includeSelfLinks}{logical}

\item{numTrials}{Integer, number of Infomap trials to minimize L}

\item{twoLevel}{logical, two level Infomap comunity detection}

\item{preclusterMultiplex}{logical}

\item{addMissingPhysicalNodes}{logical}

\item{hardPartitions}{logical}

\item{verbose}{logical}

\item{addAggregateAnalysis}{logical}

\item{multilayerRelaxRate}{Float, between 0 and 1}

\item{multilayerJSRelaxRate}{Float, between 0 and 1}

\item{outputPrefix}{String}
}
\value{
Community List
}
\description{
Return multilayer communities. INFOMAP wrapper
}
\references{
Manlio De Domenico, Andrea Lancichinetti, Alex Arenas, and Martin Rosvall, Physical Review X 5, 011027 (2015)
\href{https://doi.org/10.1103/PhysRevX.5.011027}{doi 10.1103/PhysRevX.5.011027}
\href{http://www.mapequation.org/code.html}{Source}
}
