% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultiplexityCentrality}
\alias{GetMultiplexityCentrality}
\title{Calculates Multiplexity Centrality.}
\usage{
GetMultiplexityCentrality(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
Vector of Multiplexity centralities
}
\description{
Calculates Multiplexity Centrality.
}
