% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultiplexTriadsFromNetworkList}
\alias{GetMultiplexTriadsFromNetworkList}
\title{Return the closed triangles in single layers, aggregate and multilayer network}
\usage{
GetMultiplexTriadsFromNetworkList(g.list, g.agg = NULL, verbose = T)
}
\arguments{
\item{g.list, }{a list of igraph objects encoding layers}

\item{g.agg, }{igraph object for the aggregate network}

\item{verbose}{logical}
}
\value{
A dataframe containing the closed triangles in single layers, aggregate and multilayer network
}
\description{
Return the closed triangles in single layers, aggregate and multilayer network
}
\references{
De Domenico, M., Set al. (2013). Mathematical formulation of
multilayer networks. Physical Review X, 3(4), 041022.
\href{https://doi.org/10.1103/PhysRevX.3.041022}{doi 10.1103/PhysRevX.3.041022}

Cozzo, E.,Kivela, M., De Domenico, M., et al, New J. of Phys. 17, 073029 (2015)
\href{https://doi.org/10.1088/1367-2630/17/7/073029}{doi 10.1088/1367-2630/17/7/073029}
}
