% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetMultiRWCentrality}
\alias{GetMultiRWCentrality}
\title{Returns Multi Random Walk Centrality}
\usage{
GetMultiRWCentrality(
  SupraAdjacencyMatrix,
  Layers,
  Nodes,
  Type = "classical",
  Method = "multilayer"
)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers, number of layers}

\item{Nodes}{scalar, number of nodes}

\item{Type}{string,  "classical", "pagerank", or "diffusive"}

\item{Method}{string,  "multilayer", or "perlayer"}
}
\value{
Vector of Multi-RandomWalk centralities
}
\description{
Returns Multi Random Walk Centrality
}
\references{
De Domenico, M., Solé-Ribalta, A., Omodei, E. et al. Ranking in interconnected multilayer
networks reveals versatile nodes. Nat Commun 6, 6868 (2015).
\href{https://doi.org/10.1038/ncomms7868}{doi 10.1038/ncomms7868}
}
