% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetRenyiEntropyFromAdjacencyMatrix}
\alias{GetRenyiEntropyFromAdjacencyMatrix}
\title{Calculate the quantum Renyi entropy of a network}
\usage{
GetRenyiEntropyFromAdjacencyMatrix(AdjacencyMatrix, Q = 1)
}
\arguments{
\item{AdjacencyMatrix}{an adjacency matrix}

\item{Q}{scalar, computes for the relative entropy. Computing Von Neuman quantum entropy if Q=1 (default case), and Renyi quantum entropy otherwise}
}
\value{
Renyi Entropy matrix for the provided adjacency matrix
}
\description{
Calculate the quantum Renyi entropy of a network
}
\references{
De Domenico, M., Set al. (2013). Mathematical formulation of
multilayer networks. Physical Review X, 3(4), 041022.
\href{doi 10.1103/PhysRevX.3.041022}{https://doi.org/10.1103/PhysRevX.3.041022}

De Domenico, M., et al. Structural reducibility of multilayer networks. Nat Commun 6, 6864 (2015).
\href{https://doi.org/10.1038/ncomms7864}{doi 10.1038/ncomms7864}
}
