% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetRewiredLayerWithOverlappingEdges}
\alias{GetRewiredLayerWithOverlappingEdges}
\title{Generate multiplex networks with given amount of edge overlapping}
\usage{
GetRewiredLayerWithOverlappingEdges(g, ov, tol = 0.001, verbose = FALSE)
}
\arguments{
\item{g}{graph network object}

\item{ov}{scalar in \eqn{[0, 1]} tuning the overlap}

\item{tol}{scalar, tolerance between current and desired overlap. Default 1e-3}

\item{verbose}{logical, default FALSE}
}
\value{
An igraph dataframe containing the rewiring produced
}
\description{
Generate multiplex networks with given amount of edge overlapping
}
