% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetSampleMultiplex}
\alias{GetSampleMultiplex}
\title{Get a sample of a Multiplex}
\usage{
GetSampleMultiplex(Layers, Nodes, p)
}
\arguments{
\item{Layers}{scalar, number of layers, number of layers}

\item{Nodes}{scalar, number of nodes}

\item{p}{scalar, parameter to build the sample}
}
\value{
Adjacency Matrix of the Layer and nodes indicated
}
\description{
Get a sample of a Multiplex
}
