% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{GetSPSimilarityMatrix}
\alias{GetSPSimilarityMatrix}
\title{Returns pairwise similarity based on Frobenius norm}
\usage{
GetSPSimilarityMatrix(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
Pairwise similarity based on Frobenius norm
}
\description{
Returns pairwise similarity based on Frobenius norm
}
