% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{getVersatilityProfile}
\alias{getVersatilityProfile}
\title{Return the Versatility Profile for a given centrality}
\usage{
getVersatilityProfile(mux, type)
}
\arguments{
\item{mux}{object obtained from buildMultilayerNetworkFromMuxvizFiles}

\item{type}{string, versatility type ([multi][in-|out-]degree,
[multi][in-|out-]strength, pagerank, hub, authority, multiplexity, eigenvector, katz, kcore)}
}
\value{
Vector of centralities depending the given type
}
\description{
Return the Versatility Profile for a given centrality
}
