% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_plot_functions.R
\name{layoutMultiplex}
\alias{layoutMultiplex}
\title{layoutMultiplex}
\usage{
layoutMultiplex(g.list, layout = "fr", ggplot.format = F, box = T)
}
\arguments{
\item{g.list}{list of networks}

\item{layout}{layout, as for \link[igraph]{layout_}.
Default "fr" for \link[igraph]{layout_with_fr}; other options are:
"drl", "auto", "kk", "comp", "dh".}

\item{ggplot.format}{default FALSE}

\item{box}{default TRUE}
}
\value{
a layout, i.e. a matrix of coordinates.
}
\description{
layoutMultiplex
}
