% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{modR}
\alias{modR}
\title{Return the remainder of x divided by y}
\usage{
modR(x, y)
}
\arguments{
\item{x}{numerator}

\item{y}{non-zero denominator}
}
\value{
modulo of x/y
}
\description{
Return the remainder of x divided by y
}
