% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{MultiShortestPathFromNode}
\alias{MultiShortestPathFromNode}
\title{Return the Shortest-Path from given node to other nodes, with option for counting replicates}
\usage{
MultiShortestPathFromNode(g.ext, n, Nodes, countReplicas = F)
}
\arguments{
\item{g.ext}{igraph object obtained from supra-adjacency matrix}

\item{n}{integer the node id}

\item{Nodes}{scalar, number of nodes}

\item{countReplicas}{logical}
}
\value{
A igraph object containing the list of shortest paths
}
\description{
Return the Shortest-Path from given node to other nodes, with option for counting replicates
}
