% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_plot_functions.R
\name{plot.multimodules}
\alias{plot.multimodules}
\title{Plot modules obtained from GetMultiplexCommunities_Infomap/GetMultilayerCommunities_Infomap}
\usage{
\method{plot}{multimodules}(communityList, module.colors = "auto", show.aggregate = TRUE)
}
\arguments{
\item{communityList}{list of modules}

\item{module.colors}{String}

\item{show.aggregate}{logical, default TRUE}
}
\value{
a ggplot object
}
\description{
Plot modules obtained from GetMultiplexCommunities_Infomap/GetMultilayerCommunities_Infomap
}
