% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_plot_functions.R
\name{plot.multiplex3D}
\alias{plot.multiplex3D}
\title{3D plot of a multiplex}
\usage{
\method{plot}{multiplex3D}(
  g.list,
  layer.colors,
  as.undirected = T,
  layer.layout = "auto",
  layer.labels = "auto",
  layer.labels.cex = 2,
  edge.colors = "auto",
  edge.normalize = F,
  edge.size.scale = 1,
  node.colors = "auto",
  node.size.values = 0.5,
  node.size.scale = 1,
  node.alpha = 1,
  edge.alpha = 1,
  layer.alpha = "auto",
  layout = "fr",
  show.nodeLabels = F,
  show.aggregate = F,
  aggr.alpha = "auto",
  aggr.color = "#dadada",
  node.colors.aggr = "#dadada",
  layer.scale = 2,
  layer.shift.x = 0,
  layer.shift.y = 0,
  layer.space = 1.5,
  FOV = 30
)
}
\arguments{
\item{g.list}{list of networks (representing the layers)}

\item{layer.colors}{colors of the layers (mandatory)}

\item{as.undirected}{default TRUE}

\item{layer.layout}{default "auto",}

\item{layer.labels}{default "auto",}

\item{layer.labels.cex}{default 2,}

\item{edge.colors}{default "auto",}

\item{edge.normalize}{default F,}

\item{edge.size.scale}{default 1,}

\item{node.colors}{default "auto",}

\item{node.size.values}{default 0.5,}

\item{node.size.scale}{default 1,}

\item{node.alpha}{default 1,}

\item{edge.alpha}{default 1,}

\item{layer.alpha}{default "auto",}

\item{layout}{default "fr", see \link[igraph]{layout_with_fr} for other options.}

\item{show.nodeLabels}{default F,}

\item{show.aggregate}{default F,}

\item{aggr.alpha}{default "auto",}

\item{aggr.color}{default to hex-color "#dadada",}

\item{node.colors.aggr}{default to hex-color "#dadada",}

\item{layer.scale}{default 2,}

\item{layer.shift.x}{default 0,}

\item{layer.shift.y}{default 0,}

\item{layer.space}{default 1.5,}

\item{FOV}{default 30}
}
\value{
a plot
}
\description{
3D plot of a multiplex
}
