% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_plot_functions.R
\name{plot_multimodules}
\alias{plot_multimodules}
\alias{plot.multimodules}
\title{Plot modules}
\usage{
plot_multimodules(communityList, module.colors = "auto", show.aggregate = TRUE)

plot.multimodules(communityList, module.colors = "auto", show.aggregate = TRUE)
}
\arguments{
\item{communityList}{list of modules}

\item{module.colors}{String}

\item{show.aggregate}{logical, default TRUE}
}
\value{
a ggplot object
}
\description{
Plot modules obtained from
GetMultiplexCommunities_Infomap/GetMultilayerCommunities_Infomap
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Old name
}}

