% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_plot_functions.R
\name{plot_multimotif}
\alias{plot_multimotif}
\alias{plot.multimotif}
\title{Plot multilayer motifs

Plot a motif from a motifs table returned by \code{\link{GetMultilayerMotifs}}.}
\usage{
plot_multimotif(
  motifsTable,
  motifID,
  layer.colors,
  edge.colors = "auto",
  node.colors = "auto",
  node.size = 5
)

plot.multimotif(motifsTable, motifID, layer.colors, edge.colors = "auto", 
         node.colors = "auto", node.size = 5)
}
\arguments{
\item{motifsTable}{a table as returned from \code{\link{GetMultilayerMotifs}}}

\item{motifID}{ID of the motif following the classification of \code{\link{GetMultilayerMotifs}}}

\item{layer.colors}{colors for edges/layers}

\item{edge.colors}{default "auto",}

\item{node.colors}{default "auto",}

\item{node.size}{default 5}
}
\value{
an igraph plotting object
}
\description{
Plot multilayer motifs

Plot a motif from a motifs table returned by \code{\link{GetMultilayerMotifs}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Old name
}}

