% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_plot_functions.R
\name{plot_multiplex}
\alias{plot_multiplex}
\alias{plot.multiplex}
\title{Plot a multiplex}
\usage{
plot_multiplex(
  g.list,
  layer.colors,
  edge.colors = "auto",
  node.colors = "auto",
  node.size.values = 0.5,
  node.alpha = 1,
  edge.alpha = 1,
  layout = "fr",
  show.legend = TRUE
)

plot.multiplex(g.list, layer.colors, edge.colors = "auto", node.colors = "auto",
   node.size.values = 0.5, node.alpha = 1, edge.alpha = 1, layout = "fr", show.legend = T)
}
\arguments{
\item{g.list}{list of networks (representing the layers)}

\item{layer.colors}{list of colors for the each layer}

\item{edge.colors}{edge colors. If not provided, use the same color of the layer.}

\item{node.colors}{node colors. If not provided, use the same color of the layer.}

\item{node.size.values}{default 0.5,}

\item{node.alpha}{default 1,}

\item{edge.alpha}{default 1,}

\item{layout}{default "fr", see \link[igraph]{layout_with_fr} for other options.}

\item{show.legend}{default TRUE}
}
\value{
a plot
}
\description{
Plot a multiplex
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Old name
}}

