% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{reshapeR}
\alias{reshapeR}
\title{Return a matrix reshaped according to dimension n and m}
\usage{
reshapeR(A, n, m)
}
\arguments{
\item{A}{input matrix}

\item{n}{desired width}

\item{m}{desired height}
}
\value{
Input matrix A with shape n x m
}
\description{
Return a matrix reshaped according to dimension n and m
}
