% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{ShortestPathFromNode}
\alias{ShortestPathFromNode}
\title{Return the Shortest-Path from the given node to other nodes}
\usage{
ShortestPathFromNode(g, n, Nodes)
}
\arguments{
\item{g}{an igraph object obtained from adjacency matrix}

\item{n}{integer the node id}

\item{Nodes}{scalar, number of nodes}
}
\value{
A igraph object containing the list of shortest paths
}
\description{
Return the Shortest-Path from the given node to other nodes
}
