% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{SolveEigenvalueProblem}
\alias{SolveEigenvalueProblem}
\title{Return the matrix of eigenvectors (Q) and the diagonal matrix of eigenvalues (L) as well as a vector with ordered eigenvalues (E)}
\usage{
SolveEigenvalueProblem(Matrix)
}
\arguments{
\item{Matrix}{matrix}
}
\value{
Matrix of eigenvectors (Q),
Diagonal matrix of eigenvalues (L), and vector with ordered eigenvalues (E)
}
\description{
Return the matrix of eigenvectors (Q) and the diagonal matrix of eigenvalues (L) as well as a vector with ordered eigenvalues (E)
}
