% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{speye}
\alias{speye}
\title{Return the identity matrix in sparse format}
\usage{
speye(n)
}
\arguments{
\item{n}{input matrix}
}
\value{
Identity matrix in sparse format
}
\description{
Return the identity matrix in sparse format
}
