% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{sumR}
\alias{sumR}
\title{Return the sum across rows (n=1) or columns (n=2) of a matrix}
\usage{
sumR(A, n)
}
\arguments{
\item{A}{input matrix}

\item{n}{if n=1 sum rows, if n=2 sum columns}
}
\value{
sum of rows or columns
}
\description{
Return the sum across rows (n=1) or columns (n=2) of a matrix
}
