% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{SupraAdjacencyToBlockTensor}
\alias{SupraAdjacencyToBlockTensor}
\title{Return block tensor from supra-adjacency matrix}
\usage{
SupraAdjacencyToBlockTensor(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{sparse matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
The blocks from input supra-adjacency matrix
}
\description{
Return block tensor from supra-adjacency matrix
}
