% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{SupraAdjacencyToNetworkList}
\alias{SupraAdjacencyToNetworkList}
\title{Return a list of igraph objects corresponding to layers}
\usage{
SupraAdjacencyToNetworkList(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{the supra-adajcency sparse matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
List of igraph objects corresponding to layers
}
\description{
Return a list of igraph objects corresponding to layers
}
\references{
De Domenico, M., Set al. (2013). Mathematical formulation of
multilayer networks. Physical Review X, 3(4), 041022.
\href{https://doi.org/10.1103/PhysRevX.3.041022}{doi 10.1103/PhysRevX.3.041022}
}
