% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{SupraAdjacencyToNodesTensor}
\alias{SupraAdjacencyToNodesTensor}
\title{Return nodes tensor from supra-adjacency matrix}
\usage{
SupraAdjacencyToNodesTensor(SupraAdjacencyMatrix, Layers, Nodes)
}
\arguments{
\item{SupraAdjacencyMatrix}{sparse matrix}

\item{Layers}{scalar, number of layers}

\item{Nodes}{scalar, number of nodes}
}
\value{
The diagonal blocks from a supra-adajcency matrix.
}
\description{
Return nodes tensor from supra-adjacency matrix
}
