% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{traceR}
\alias{traceR}
\title{Return the trace of input matrix}
\usage{
traceR(A)
}
\arguments{
\item{A}{input matrix}
}
\value{
Trace of input matrix
}
\description{
Return the trace of input matrix
}
