% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_annotated.R
\name{binarizeMatrix}
\alias{binarizeMatrix}
\title{Gets a binarized version (only 0s and 1s) of a Matrix, which is assumed to be sparse}
\usage{
binarizeMatrix(A)
}
\arguments{
\item{A}{matrix}
}
\value{
Matrix binarized
}
\description{
Gets a binarized version (only 0s and 1s) of a Matrix, which is assumed to be sparse
}
