% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muxLib_plot_functions.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{plots to to be arranged in the multiplot, alternatively provide a list \code{plotlist}}

\item{plotlist}{list of plots to be arranged}

\item{file}{a file}

\item{cols}{number of columns}

\item{layout}{an igraph layout}
}
\value{
a plot
}
\description{
Multiplot
}
